requirements.txt
django==1.11.20
sqlalchemy==0.8.7
setuptools==40.7.1
MySQL-python==1.2.5
python-memcached==1.59
python-ldap==3.1.0
pyodbc==4.0.25
informixdb==2.5
bootstrap_admin==0.4.0
pysftp==0.2.9
requests==2.21.0
lxml==4.3.1
reportlab==3.5.13
python-dateutil==2.8.0
smartystreets-python-sdk==4.0.1
pytz==2018.9
awscli==1.16.152
pycrypto==2.6.1
